pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "uptrade"
version = "0.1.0"
description = "AI-Powered Crypto Trading Platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "vectorbtpro",
    "numba>=0.59",
    "numpy>=1.24",
    "pandas>=2.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "polygon-api-client>=1.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.30",
    "plotly>=5.18",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
